% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blueprint_list}
\alias{blueprint_list}
\title{A tamRgo Pet Blueprint (List)}
\format{
A list with two elements of length 5 and 3 respectively, containing
character or integer vectors of length 1.
\describe{
\item{characteristics}{'name' (character), 'species' (character), 'stage' (integer), 'born' (character) and 'age' (integer)}
\item{status}{'hungry' (integer), 'happy' (integer) and 'dirty' (integer)}
}
}
\usage{
blueprint_list
}
\description{
A demo blueprint of a tamRgo pet as a list.
}
\keyword{datasets}
