% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blueprint_yaml}
\alias{blueprint_yaml}
\title{A tamRgo Pet Blueprint (YAML String)}
\format{
A character string that represents a YAML file with two lists of 5
and 3 members respectively.
\describe{
\item{characteristics}{'name' (character), 'species' (character), 'stage' (integer), 'born' (character) and 'age' (integer)}
\item{status}{'hungry' (integer), 'happy' (integer) and 'dirty' (integer)}
}
}
\usage{
blueprint_yaml
}
\description{
A demo blueprint of a tamRgo pet as a character string that represents a YAML
file.
}
\keyword{datasets}
