% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{clean}
\alias{clean}
\title{Clean Your Pet}
\usage{
clean()
}
\value{
Nothing.
}
\description{
Clean dirt off your pet, which reduces your pet's 'dirty'
status value by 1, down to a minimum of 0.
}
\examples{
\dontrun{clean()}
}
