% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{clean_pet}
\alias{clean_pet}
\title{Clean A tamRgo Pet}
\usage{
clean_pet(pet_id = Sys.getenv("TAMRGO_PET_ID"))
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
stored in the user's Renviron.}
}
\value{
Nothing. Messages are printed to the console and a GitHub gist may be
updated.
}
\description{
Clean a tamRgo pet, decreasing its 'dirty' status by 1.
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
clean_pet(pet_id = gist_id)
}
}
