% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_blueprint}
\alias{create_blueprint}
\title{Generate A Blueprint For A New tamRgo Pet}
\usage{
create_blueprint(pet_name, seed = NULL)
}
\arguments{
\item{pet_name}{Character. A name for hte new tamRgo pet.}

\item{seed}{Integer or NULL. Supply a seed value to recreate a given tamRgo
pet's sampled values. Defaults to NULL.}
}
\value{
A list of lists. See details.
}
\description{
Generate A Blueprint For A New tamRgo Pet
}
\details{
A tamRgo 'blueprint' is a list of lists. The sublist 'characteristics'
contains:
\describe{
\item{species}{Placeholder}
\item{stage}{Placeholder}
\item{born}{Placeholder}
\item{age}{Placeholder}
}
The sublist 'status' contains
\describe{
\item{hunger}{Placeholder}
\item{happy}{Placeholder}
\item{clean}{Placeholder}
}
}
\examples{
\dontrun{create_blueprint("Kevin", 1234)}
}
