% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-status.R
\name{.change_status}
\alias{.change_status}
\title{Change A Blueprint Status Value}
\usage{
.change_status(
  pet_id = Sys.getenv("TAMRGO_PET_ID"),
  status_type = c("happy", "hungry", "dirty"),
  change = 1L,
  min_value = 0L,
  max_value = 5L
)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
stored in the user's Renviron.}

\item{status_type}{Character. The status that will be updated. One of
'hungry', 'happy' or 'dirty'.}

\item{change}{Integer. The amount that the status value should change by (can
be positive or negative, but must be between the values of min_value and
max_value)}

\item{min_value}{Integer. The minimum value that the status value can take.}

\item{max_value}{Integer. The maximum value that the status value can take.}
}
\value{
Nothing. Messages may be printed to the console and a GitHub gist may
be updated.
}
\description{
Change the status value of a pet. This may happen over time (i.e. hunger
should increase) or as a result of player interaction (e.g. feeding will
reduce hunger).
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
.change_status(pet_id = gist_id, status_type = "happy", change = -1L)
}
}
