% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-environ.R
\name{.check_environ_exists}
\alias{.check_environ_exists}
\title{Check the 'tamRgo.environ' File Exists in User's Home Directory}
\usage{
.check_environ_exists(
  env_path = file.path(Sys.getenv("HOME"), "tamRgo.environ")
)
}
\value{
Logical.
}
\description{
Check that the tamRgo.environ file exisst in the user's home directory. This
is the persistent file that contains a pet ID (i.e. gist ID) stored as
TAMRGO_PET_ID.
}
\examples{
\dontrun{.check_environ_exists()}
}
