% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints.R
\name{.create_blueprint}
\alias{.create_blueprint}
\title{Create a Blueprint}
\usage{
.create_blueprint(pet_name)
}
\arguments{
\item{pet_name}{Character. A name for the new tamRgo pet. Maximum eight
characters.}
}
\value{
A list.
}
\description{
Create a Blueprint
}
\examples{
\dontrun{.create_blueprint(name = "Kevin")}
}
