% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renviron.R
\name{.create_environ}
\alias{.create_environ}
\title{Create an 'tamRgo.environ' File in User's Home Directory}
\usage{
.create_environ(pet_id, read = TRUE)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
stored in the user's Renviron.}

\item{read}{Logical. Read back the value from the newly-created
tamRgo.environ file in the Renviron? Defaults to TRUE}
}
\value{
Nothing.
}
\description{
Create a persistent file in the user's home directory called tamRgo.environ
and write to it a provided pet ID (i.e. gist ID) value, with the label
TAMRGO_PET_ID.
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
.create_environ(pet_id = gist_id)
}
}
