% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-environ.R
\name{.delete_environ}
\alias{.delete_environ}
\title{Delete the 'tamRgo.environ' File in User's Home Directory}
\usage{
.delete_environ()
}
\value{
Nothing.
}
\description{
Delete the tamRgo.environ file in the user's home directory. This is the
persistent file that contains a pet ID (i.e. gist ID) stored as
TAMRGO_PET_ID.
}
\examples{
\dontrun{.delete_environ()}
}
