% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renviron.R
\name{.destroy_environ}
\alias{.destroy_environ}
\title{Delete the 'tamRgo.environ' File in User's Home Directory}
\usage{
.destroy_environ()
}
\value{
Nothing.
}
\description{
Delete the tamRgo.environ file in the user's home directory. This is the
persistent file that contains a pet ID (i.e. gist ID) stored as
TAMRGO_PET_ID.
}
\examples{
\dontrun{.destroy_environ(pet_id = gist_id)}
}
