% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gists.R
\name{.patch_blueprint}
\alias{.patch_blueprint}
\title{Update A tamRgo Pet's Blueprint}
\usage{
.patch_blueprint(
  pet_id = Sys.getenv("TAMRGO_PET_ID"),
  what = c("pet_id", "name", "species", "born", "stage", "age", "xp", "hungry",
    "happy", "dirty"),
  new_value
)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
that's in the current environment.}

\item{what}{Character. The name of the characteristic or status to be updated
in the pet blueprint. See details.}

\item{new_value}{Character or integer (see details). A value to overwrite the
current value in the blueprint (GitHub gist YAML) for the blueprint
variable provided by the 'what' argument.}
}
\value{
Nothing.
}
\description{
Update an existing GitHub gist that contains a tamRgo pet's 'blueprint' (its
list of characteristics and status values).
}
\details{
A tamRgo 'blueprint' is a list of two lists ('characteristics' and 'status')
that stores information about a pet. The sublist 'characteristics' contains:
\describe{
\item{pet_id}{Unique pet identification number (GitHub Gist ID number)}
\item{name}{The pet's name}
\item{species}{Type of pet}
\item{born}{Date of birth}
\item{stage}{Growth stage reached}
\item{age}{Days since birth}
\item{xp}{Experience points}
}
The sublist 'status' contains:
\describe{
\item{hungry}{Hunger on a scale of 1 (least) to 5 (most)}
\item{happy}{Happiness on a scale of 1 (least) to 5 (most)}
\item{dirty}{Dirtiness on a scale of 1 (least) to 5 (most)}
}
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
.patch_blueprint(pet_id = gist_id)
}
}
