% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gists.R
\name{.post_blueprint}
\alias{.post_blueprint}
\title{Create A tamRgo Pet's Blueprint}
\usage{
.post_blueprint(blueprint)
}
\arguments{
\item{blueprint}{A list of lists. Contains characteristics and status
values for a given tamRgo pet. See details.}
}
\value{
Nothing.
}
\description{
Write a new GitHub gist with the name 'tamRgo.yaml' that contains a tamRgo
pet's 'blueprint' (its list of characteristics and status values).
}
\details{
A tamRgo 'blueprint' is a list of two lists ('characteristics' and 'status')
that stores information about a pet. The sublist 'characteristics' contains:
\describe{
\item{species}{Type of pet (charcter)}
\item{stage}{Growth stage reached (integer)}
\item{born}{Date of birth (character)}
\item{age}{Days since birth (integer)}
}
The sublist 'status' contains:
\describe{
\item{hungry}{Hunger (integer, 1 to 5 where 5 is most hungry)}
\item{happy}{Happiness (integer, 1 to 5 where 5 is most happy)}
\item{dirty}{Dirtiness (integer, 1 to 5 where 5 is most dirty)}
}
}
\examples{
\dontrun{
bp <- create_blueprint(name = "Kevin", seed = 1234L)
.post_blueprint(blueprint = bp)
}
}
