% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-read.R
\name{.read_blueprint}
\alias{.read_blueprint}
\title{Read A tamRgo Pet's Blueprint}
\usage{
.read_blueprint(pet_id)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint.}
}
\value{
A list. See details.
}
\description{
Read a YAML file from a GitHub Gist that contains a tamRgo 'blueprint'.
}
\details{
A tamRgo 'blueprint' is a list of two lists ('characteristics' and 'status')
that stores information about a pet. The sublist 'characteristics' contains:
\describe{
\item{species}{Type of pet}
\item{stage}{Growth stage reached}
\item{born}{Date of birth}
\item{age}{Days since birth}
}
The sublist 'status' contains:
\describe{
\item{hungry}{Hunger on a scale of 1 (least) to 5 (most)}
\item{happy}{Happiness on a scale of 1 (least) to 5 (most)?}
\item{dirty}{Dirtiness on a scale of 1 (least) to 5 (most)?}
}
}
\examples{
\dontrun{read_blueprint("9a0138f474f15ceabf267e704c830625")}
}
