% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-environ.R
\name{.read_environ}
\alias{.read_environ}
\title{Read the 'tamRgo.environ' File in User's Home Directory}
\usage{
.read_environ(env_path = file.path(Sys.getenv("HOME"), "tamRgo.environ"))
}
\arguments{
\item{env_path}{Character. Path to a possible 'tamRgo.environ' file on the
user's machine.}
}
\value{
Nothing.
}
\description{
Read into the environment a pet ID (i.e. gist ID) that's stored as
TAMRGO_PET_ID in the tamRgo.environ file in the user's home directory,
}
\examples{
\dontrun{.read_environ()}
}
