% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints.R
\name{.roll_characteristics}
\alias{.roll_characteristics}
\title{Sample Characteristics for a New Pet}
\usage{
.roll_characteristics()
}
\value{
A list.
}
\description{
Sample Characteristics for a New Pet
}
\examples{
\dontrun{.roll_characteristics()}
}
