% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renviron.R
\name{.set_renviron}
\alias{.set_renviron}
\title{Set A tamRgo Pet ID In The Renviron}
\usage{
.set_renviron(pet_id = Sys.getenv("TAMRGO_PET_ID"), overwrite = TRUE)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value stored
in the user's Renviron.}

\item{overwrite}{Logical. Should an existing pet ID value in the Renviron be
overwritten by the supplied pet_id? Defaults to TRUE.}
}
\value{
Nothing. Updates the user's Renviron file, prints messages to the
console.
}
\description{
Set the pet ID (i.e. GitHub gist ID) as the 'TAMRGO_PET_ID' value in the
user's Renviron file. This prevents a user having to supply the pet_id
whenever they interact with their pet.
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
.set_renviron(pet_id = gist_id)
}
}
