% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renviron.R
\name{.unset_renviron}
\alias{.unset_renviron}
\title{Unset A tamRgo Pet ID From The Renviron}
\usage{
.unset_renviron(pet_id = Sys.getenv("TAMRGO_PET_ID"))
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value stored
in the user's Renviron.}
}
\value{
Nothing. Updates the user's Renviron file, prints messages to the
console.
}
\description{
Set the pet ID (i.e. GitHub gist ID) as the 'TAMRGO_PET_ID' value in the
user's Renviron file. This prevents a user having to supply the pet_id
whenever they interact with their pet.
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
.unset_renviron(pet_id = gist_id)
}
}
