% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints.R
\name{.update_blueprint}
\alias{.update_blueprint}
\title{Update Blueprint Given Time}
\usage{
.update_blueprint(
  happy_increment = 10L,
  hungry_increment = 15L,
  dirty_increment = 30L,
  xp_increment = 5L
)
}
\arguments{
\item{happy_increment}{Integer. How many minutes must elapse before the
'happy' status value decreases by 1?}

\item{hungry_increment}{Integer. How many minutes must elapse before the
'hungry' status value decreases by 1?}

\item{dirty_increment}{Integer. How many minutes must elapse before the
'dirty' status value decreases by 1?}

\item{xp_increment}{Integer. How many minutes must elapse before the pet
gains 1 XP (experience point)?}
}
\value{
A list.
}
\description{
Update time-dependent blueprint values given how much time has
elapsed since the last recorded interaction.
}
\examples{
\dontrun{.update_blueprint()}
}
