% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints.R
\name{.write_blueprint}
\alias{.write_blueprint}
\title{Write a Local Blueprint}
\usage{
.write_blueprint(blueprint, ask = TRUE)
}
\arguments{
\item{blueprint}{List. A pet's blueprint.}

\item{ask}{Logical. Should the user be asked about creating or updating the
existing blueprint file? Defaults to TRUE.}
}
\value{
Nothing.
}
\description{
Write a Local Blueprint
}
\examples{
\dontrun{
    bp <- .create_blueprint("KEVIN")
    .write_blueprint(bp, ask = FALSE)
    }
}
