% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renviron.R
\name{.write_renviron}
\alias{.write_renviron}
\title{Write A Pet ID To An Renviron File}
\usage{
.write_renviron(pet_id, overwrite = FALSE)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint.}

\item{overwrite}{Logical. Should an existing pet ID value in the Renviron be
overwritten by the supplied pet_id?}
}
\value{
Nothing. Updates the user's Renviron file, prints messages to the
console.
}
\description{
Set the pet ID (i.e. GitHub gist ID) as the 'TAMRGO_PET_ID' value in the
user's Renviron file. This prevents a user having to supply the pet_id
whenever they interact with their pet.
}
\examples{
\dontrun{.write_id_to_renviron("1234567890")}
}
