% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{feed}
\alias{feed}
\title{Feed Your Pet}
\usage{
feed()
}
\value{
Nothing.
}
\description{
Give food to your pet, which reduces your pet's 'hungry' status
value by 1, down to a minimum of 0.
}
\examples{
\dontrun{feed()}
}
