% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{get_stats}
\alias{get_stats}
\title{Print Pet Statistics}
\usage{
get_stats()
}
\value{
Nothing.
}
\description{
Print to the console your pet's current characteristics and
status values.
}
\details{
The output will show the following elements:

\describe{
\item{Name}{Pet's user-provided name.}
\item{Species}{Randomly-selected pet species.}
\item{Age}{Days since born.}
\item{Level}{Growth stage.}
\item{Alive}{Is the pet alive?}
\item{Happy}{Happiness on a scale of 0 to 5.}
\item{Hungry}{Hunger on a scale of 0 to 5.}
\item{Dirty}{Dirtiness on a scale of 0 to 5.}
}
}
\examples{
\dontrun{get_stats()}
}
