% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{load_pet}
\alias{load_pet}
\title{Load A tamRgo Pet}
\usage{
load_pet(pet_id = Sys.getenv("TAMRGO_PET_ID"))
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
stored in the user's Renviron.}
}
\value{
Nothing.
}
\description{
Fetch a pet ID (i.e. GitHub gist ID) from the TAMRGO_PET_ID variable in the
tamRgo.environ file on the user's home directory. If tamRgo.environ doesn't
exist, then prompt to create it.
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
load_pet(pet_id = gist_id)
}
}
