% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{play}
\alias{play}
\title{Play with Your Pet}
\usage{
play()
}
\value{
Nothing.
}
\description{
Play a game of chance with your pet, which increases your pet's
'happy' status value by 1, up to a maximum of 5.
}
\examples{
\dontrun{play()}
}
