% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{play_with_pet}
\alias{play_with_pet}
\title{Play With A tamRgo Pet}
\usage{
play_with_pet(pet_id = Sys.getenv("TAMRGO_PET_ID"))
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
stored in the user's Renviron.}
}
\value{
Nothing.
}
\description{
Play with a tamRgo pet, increasing its 'happy' status by 1.
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
play_with_pet(pet_id = gist_id)
}
}
