% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{post_blueprint}
\alias{post_blueprint}
\title{Create A tamRgo Pet's Blueprint}
\usage{
post_blueprint(tamRgo_blueprint)
}
\arguments{
\item{tamRgo_blueprint}{A list of lists. Contains characteristics and status
values for a given tamRgo pet. See details.}
}
\value{
Nothing.
}
\description{
Write a new tamRgo pet's 'blueprint' (a list of characteristics and status
values) to a new GitHub gist.
}
\details{
A tamRgo 'blueprint' is a list of lists. The sublist 'characteristics'
contains:
\describe{
\item{species}{Placeholder}
\item{stage}{Placeholder}
\item{born}{Placeholder}
\item{age}{Placeholder}
}
The sublist 'status' contains
\describe{
\item{hunger}{Placeholder}
\item{happy}{Placeholder}
\item{clean}{Placeholder}
}
}
\examples{
\dontrun{post_blueprint(blueprint)}
}
