% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{read_blueprint}
\alias{read_blueprint}
\title{Read A tamRgo Pet's Blueprint From A GitHub Gist}
\usage{
read_blueprint(tamRgo_id)
}
\arguments{
\item{tamRgo_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint.}
}
\value{
A list. See details.
}
\description{
Read a YAML file from a GitHub Gist that contains a tamRgo 'blueprint'.
}
\details{
A tamRgo 'blueprint' is a list of lists. The sublist 'characteristics'
contains:
\describe{
\item{species}{Placeholder}
\item{stage}{Placeholder}
\item{born}{Placeholder}
\item{age}{Placeholder}
}
The sublist 'status' contains
\describe{
\item{hunger}{Placeholder}
\item{happy}{Placeholder}
\item{clean}{Placeholder}
}
}
\examples{
\dontrun{read_blueprint("9a0138f474f15ceabf267e704c830625")}
}
