% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{release_pet}
\alias{release_pet}
\title{Release A tamRgo Pet Forever}
\usage{
release_pet(pet_id = Sys.getenv("TAMRGO_PET_ID"))
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
stored in the user's Renviron.}
}
\value{
Nothing. Interactive messages are printed to the console and a GitHub
gist may be deleted.
}
\description{
Delete a tamRgo pet's blueprint (YAML file) from its associated GitHub gist.
Also unsets the pet ID (TAMRGO_PET_ID) from the Renviron.
}
\examples{
\dontrun{release_pet()}
}
