% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{see_pet}
\alias{see_pet}
\title{See Your Pet}
\usage{
see_pet()
}
\value{
Nothing.
}
\description{
Print to the console an image of your digital pet.
}
\details{
The appearance of your pet is dependent on its species and level,
which you can view with \code{\link{get_stats}}.
}
\examples{
\dontrun{see_pet()}
}
