% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{see_stats}
\alias{see_stats}
\title{Print Your Pet's Stats}
\usage{
see_stats()
}
\value{
Nothing.
}
\description{
Print to the console the characteristics, experience and status
of the current pet.
}
\examples{
\dontrun{see_stats()}
}
