% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamRgo-package.R
\docType{package}
\name{tamRgo-package}
\alias{tamRgo}
\alias{tamRgo-package}
\title{tamRgo: Digital Pets in R}
\description{
Create and care for a digital pet inside R sessions.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

Other contributors:
\itemize{
  \item Adriana De Palma [contributor]
}

}
\keyword{internal}
