% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{update_blueprint}
\alias{update_blueprint}
\title{Update A tamRgo Pet's Blueprint From An Interactive Session.}
\usage{
update_blueprint(tamRgo_id)
}
\arguments{
\item{tamRgo_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint.}
}
\value{
Nothing.
}
\description{
Update an existing tamRgo pet's 'blueprint' (its list of characteristics and
status values) in an existing GitHub gist.
}
\examples{
\dontrun{update_blueprint("9a0138f474f15ceabf267e704c830625")}
}
