% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-blueprints.R
\name{.create_blueprint}
\alias{.create_blueprint}
\title{Generate A Blueprint For A New tamRgo Pet}
\usage{
.create_blueprint(pet_name, seed = NULL)
}
\arguments{
\item{pet_name}{Character. A name for the new tamRgo pet.}

\item{seed}{Integer or NULL. Supply a seed value to recreate a given tamRgo
pet's sampled values. Defaults to NULL.}
}
\value{
A list. See details.
}
\description{
Generate A Blueprint For A New tamRgo Pet
}
\details{
A tamRgo 'blueprint' is a list of two lists ('characteristics' and 'status')
that stores information about a pet. The sublist 'characteristics' contains:
\describe{
\item{species}{Type of pet}
\item{stage}{Growth stage reached}
\item{born}{Date of birth}
\item{age}{Days since birth}
}
The sublist 'status' contains:
\describe{
\item{hungry}{Hunger on a scale of 1 (least) to 5 (most)}
\item{happy}{Happiness on a scale of 1 (least) to 5 (most)}
\item{dirty}{Dirtiness on a scale of 1 (least) to 5 (most)}
}
}
\examples{
\dontrun{.create_blueprint(name = "Kevin", seed = 1234L)}
}
