% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gists.R
\name{.get_blueprint}
\alias{.get_blueprint}
\title{Read A tamRgo Pet's Blueprint}
\usage{
.get_blueprint(pet_id = Sys.getenv("TAMRGO_PET_ID"))
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
stored in the user's Renviron.}
}
\value{
A list. See details.
}
\description{
Read a YAML file from a GitHub Gist that contains a tamRgo 'blueprint'.
}
\details{
A tamRgo 'blueprint' is a list of two lists ('characteristics' and 'status')
that stores information about a pet. The sublist 'characteristics' contains:
\describe{
\item{species}{Type of pet (charcter)}
\item{stage}{Growth stage reached (integer)}
\item{born}{Date of birth (character)}
\item{age}{Days since birth (integer)}
}
The sublist 'status' contains:
\describe{
\item{hungry}{Hunger (integer, 1 to 5 where 5 is most hungry)}
\item{happy}{Happiness (integer, 1 to 5 where 5 is most happy)}
\item{dirty}{Dirtiness (integer, 1 to 5 where 5 is most dirty)}
}
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
.get_blueprint(pet_id = gist_id)
}
}
