% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gists.R
\name{.post_blueprint}
\alias{.post_blueprint}
\title{Create A tamRgo Pet's Blueprint}
\usage{
.post_blueprint(blueprint)
}
\arguments{
\item{blueprint}{A list of lists. Contains characteristics and status
values for a given tamRgo pet. See details.}
}
\value{
Nothing.
}
\description{
Write a new GitHub gist with the name 'tamRgo.yaml' that contains a tamRgo
pet's 'blueprint' (its list of characteristics and status values).
}
\details{
A tamRgo 'blueprint' is a list of two lists ('characteristics' and 'status')
that stores information about a pet. The sublist 'characteristics' contains:
\describe{
\item{pet_id}{Unique pet identification number (GitHub Gist ID number)}
\item{name}{The pet's name}
\item{species}{Type of pet}
\item{born}{Date of birth}
\item{stage}{Growth stage reached}
\item{age}{Days since birth}
\item{xp}{Experience points}
}
The sublist 'status' contains:
\describe{
\item{hungry}{Hunger on a scale of 1 (least) to 5 (most)}
\item{happy}{Happiness on a scale of 1 (least) to 5 (most)}
\item{dirty}{Dirtiness on a scale of 1 (least) to 5 (most)}
}
}
\examples{
\dontrun{
bp <- create_blueprint(name = "Kevin", seed = 1234L)
.post_blueprint(blueprint = bp)
}
}
