% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprints.R
\name{.read_blueprint}
\alias{.read_blueprint}
\title{Read a Local Blueprint}
\usage{
.read_blueprint()
}
\value{
Nothing.
}
\description{
Read a Local Blueprint
}
\examples{
\dontrun{.create_blueprint(name = "Kevin")}
}
