% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-blueprints.R
\name{.roll_characteristics}
\alias{.roll_characteristics}
\title{Sample Characteristics For A New tamRgo Pet}
\usage{
.roll_characteristics(seed = NULL)
}
\arguments{
\item{seed}{Integer or NULL. Supply a seed value to recreate a given tamRgo
pet's sampled values. Defaults to NULL.}
}
\value{
A list.
}
\description{
Sample Characteristics For A New tamRgo Pet
}
\examples{
\dontrun{.roll_characteristics(seed = 1234L)}
}
