% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-write.R
\name{.update_blueprint}
\alias{.update_blueprint}
\title{Update A tamRgo Pet's Blueprint}
\usage{
.update_blueprint(pet_id)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint.}
}
\value{
Nothing.
}
\description{
Update an existing GitHub gist that contains a tamRgo pet's 'blueprint' (its
list of characteristics and status values).
}
\examples{
\dontrun{update_blueprint("9a0138f474f15ceabf267e704c830625")}
}
