% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gists.R
\name{.update_blueprint}
\alias{.update_blueprint}
\title{Update A tamRgo Pet's Blueprint}
\usage{
.update_blueprint(pet_id = Sys.getenv("TAMRGO_PET_ID"))
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value stored
in the user's Renviron.}
}
\value{
Nothing.
}
\description{
Update an existing GitHub gist that contains a tamRgo pet's 'blueprint' (its
list of characteristics and status values).
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
.update_blueprint(pet_id = gist_id)
}
}
