% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gists.R
\name{.update_blueprint}
\alias{.update_blueprint}
\title{Update A tamRgo Pet's Blueprint}
\usage{
.update_blueprint(
  pet_id = Sys.getenv("TAMRGO_PET_ID"),
  what = c("name", "species", "stage", "born", "age", "hungry", "happy", "dirty"),
  new_value
)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value stored
in the user's Renviron.}

\item{what}{Character. The name of the characteristic or status to be updated
in the pet blueprint. See details.}

\item{new_value}{Character or integer (see details). A value to overwrite the
current value in the blueprint (GitHub gist YAML) for the blueprint
variable provided by the 'what' argument.}
}
\value{
Nothing.
}
\description{
Update an existing GitHub gist that contains a tamRgo pet's 'blueprint' (its
list of characteristics and status values).
}
\details{
A tamRgo 'blueprint' is a list of two lists ('characteristics' and 'status')
that stores information about a pet. The sublist 'characteristics' contains:
\describe{
\item{species}{Type of pet (charcter)}
\item{stage}{Growth stage reached (integer)}
\item{born}{Date of birth (character)}
\item{age}{Days since birth (integer)}
}
The sublist 'status' contains:
\describe{
\item{hungry}{Hunger (integer, 1 to 5 where 5 is most hungry)}
\item{happy}{Happiness (integer, 1 to 5 where 5 is most happy)}
\item{dirty}{Dirtiness (integer, 1 to 5 where 5 is most dirty)}
}
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
.update_blueprint(pet_id = gist_id)
}
}
