% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{get_stats}
\alias{get_stats}
\title{Print Your Pet's Stats}
\usage{
get_stats()
}
\value{
Nothing.
}
\description{
Print to the console the characteristics, experience and status
of the current pet.
}
\details{
The  output will show the following elements of the blueprint:

Section 'characteristics':
\describe{
\item{name}{Pet's user-provided name.}
\item{species}{Randomly-selected pet species.}
\item{born}{Date that the pet was created.}
\item{age}{Days since born.}
\item{level}{Growth stage.}
\item{xp}{Experience points.}
\item{happy}{Happiness on a scale of 0 to 5.}
\item{hungry}{Hunger on a scale of 0 to 5.}
\item{dirty}{Dirtiness on a scale of 0 to 5.}
}
}
\examples{
\dontrun{see_stats()}
}
