% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{lay_egg}
\alias{lay_egg}
\title{Generate A New tamRgo Pet}
\usage{
lay_egg(pet_name)
}
\arguments{
\item{pet_name}{Character. A name for the new tamRgo pet.}
}
\value{
Nothing. A new GitHub gist is created and a message printed ot the
console.
}
\description{
Create a new tamRgo pet blueprint and write it to a new gist
}
\examples{
\dontrun{lay_egg("Kevin")}
}
