% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{lay_egg}
\alias{lay_egg}
\title{Create a New Pet}
\usage{
lay_egg(pet_name)
}
\arguments{
\item{pet_name}{Character. A name for the new tamRgo pet. Maximum eight
characters.}
}
\value{
Nothing.
}
\description{
Generate and save a new pet's blueprint. The pet begins life as
an egg.
}
\examples{
\dontrun{lay_egg(name = "KEVIN")}
}
