% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{lay_egg}
\alias{lay_egg}
\title{Generate a New Pet}
\usage{
lay_egg(pet_name)
}
\arguments{
\item{pet_name}{Character. A name for your new pet. Maximum eight characters.}
}
\value{
Nothing.
}
\description{
Lays an egg that contains a new digital pet who will live on
your computer.
}
\details{
A persistent 'blueprint' file of your pet's characteristics will be
saved to your computer. It will be saved as an RDS to the directory
location given by \code{tools::R_user_dir("tamRgo", which = "data")}. You can
only have store one blueprint at a time, so you can only have one pet at
a time on your computer. .
}
\examples{
\dontrun{lay_egg(pet_name = "KEVIN")}
}
