% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{load_pet}
\alias{load_pet}
\title{Load A tamRgo Pet}
\usage{
load_pet(pet_id, overwrite = TRUE)
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a gist that contains a given
tamRgo pet's blueprint.}

\item{overwrite}{Logical. Should an existing pet ID value in the Renviron be
overwritten by the supplied pet_id? Defaults to TRUE.}
}
\value{
Nothing. Messages are printed to the console and an Renviron variable
is updated.
}
\description{
Set a pet_id (i.e. GitHub gist ID) as the Renviron variable TAMRGO_PET_ID.
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
load_pet(pet_id = gist_id)
}
}
