% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{see_stats}
\alias{see_stats}
\title{See A Stats Summary For A tamRgo Pet}
\usage{
see_stats(pet_id = Sys.getenv("TAMRGO_PET_ID"))
}
\arguments{
\item{pet_id}{Character. A GitHub gist ID for a YAML file that contains a
given tamRgo pet's blueprint. By default it uses the TAMRGO_PET_ID value
stored in the user's Renviron.}
}
\value{
Nothing. A message is printed to the console.
}
\description{
Print to the console the current characteristics and status of a tamRgo pet.
}
\examples{
\dontrun{
gist_id <- "1234567890abcdefghijklmnopqrstuv"
see_stats(pet_id = gist_id)
}
}
