% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamRgo-package.R
\docType{package}
\name{tamRgo-package}
\alias{tamRgo}
\alias{tamRgo-package}
\title{tamRgo: Digital Pets for R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Store a persistent digital pet on your computer and interact with it in your R console.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matt-dray.github.io/tamRgo/}
  \item \url{https://github.com/matt-dray/tamRgo}
  \item Report bugs at \url{https://github.com/matt-dray/tamRgo/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

Other contributors:
\itemize{
  \item Adriana De Palma [contributor]
}

}
\keyword{internal}
