% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamRgo-package.R
\docType{package}
\name{tamRgo-package}
\alias{tamRgo}
\alias{tamRgo-package}
\title{tamRgo: Digital Pets For R}
\description{
Create and care for a digital pet, using GitHub gists to store pet 'blueprints'.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
